+++ src/main/java/com/avaliacao/crud/service/DocumentoService.java
@@ -1,7 +1,9 @@
 package com.avaliacao.crud.service;
 
 import com.avaliacao.crud.dto.DocumentoRequestDTO;
+import com.avaliacao.crud.model.Beneficiario;
 import com.avaliacao.crud.model.Documento;
+import com.avaliacao.crud.repository.BeneficiarioRepository;
 import com.avaliacao.crud.repository.DocumentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,9 @@ public class DocumentoService {
     @Autowired
     private DocumentoRepository documentoRepository;
 
+    @Autowired
+    private BeneficiarioRepository beneficiarioRepository;
+
     public List<Documento> listAllDocuments(){
         return documentoRepository.findAll();
     }
@@ -24,7 +29,7 @@ public class DocumentoService {
         return documentoRepository.findByIdBeneficiario(idBeneficiario);
     }
 
-    public Documento addDocumento(DocumentoRequestDTO documentoRequestDTO) {
+    public void addDocumento(DocumentoRequestDTO documentoRequestDTO) throws Exception {
 
         Documento documento = new Documento();
 
@@ -33,24 +38,39 @@ public class DocumentoService {
         documento.setDataInclusao(LocalDateTime.now());
         documento.setIdBeneficiario(documentoRequestDTO.getIdBeneficiario());
 
-        return documentoRepository.save(documento);
+        Optional<Beneficiario> beneficiario = beneficiarioRepository.findById(documentoRequestDTO.getIdBeneficiario());
+
+        if(!beneficiario.isPresent()){
+            throw new Exception("Nenhum beneficiário encontrado com esse id");
+        }
+
+        documentoRepository.save(documento);
     }
 
-    public Documento editDocumento(Long id, DocumentoRequestDTO documentoRequestDTO){
+    public void editDocumento(Long id, DocumentoRequestDTO documentoRequestDTO) throws Exception {
 
         Optional<Documento> documento = documentoRepository.findById(id);
+
+        if(!documento.isPresent()){
+            throw new Exception("Nenhum documento encontrado com esse id");
+        }
 
         documento.get().setTipoDocumento(documentoRequestDTO.getTipoDocumento());
         documento.get().setDescricao(documentoRequestDTO.getDescricao());
         documento.get().setDataAtualizacao(LocalDateTime.now());
         documento.get().setIdBeneficiario(documentoRequestDTO.getIdBeneficiario());
 
-        return documentoRepository.save(documento.get());
+        documentoRepository.save(documento.get());
     }
 
-    public void deleteDocumento(Long id){
+    public void deleteDocumento(Long id) throws Exception {
 
         Optional<Documento> documento = documentoRepository.findById(id);
+
+        if(!documento.isPresent()){
+            throw new Exception("Nenhum documento encontrado com esse id");
+        }
+
         documentoRepository.delete(documento.get());
     }
